+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- ui: implement AppGeneralSettingsModal
 ### Changed
 - ui: rework brightness controls in DisplayCard, sync brightness state after turning on auto-brightness